get zap invoices in useZaps hook
